refactor(logger): do not expose internals
BREAKING CHANGE:
- the LoggerFeature type and LoggerFeatureKind enum are no longer exported
