+++ libs/core/src/lib/logger/console/with-browser-console-transport.function.ts
@@ -1,8 +1,9 @@
-import { LoggerFeature, LoggerFeatureKind } from '../provide-logger'
 import { ConsoleLogTransportConfig } from './console-log-transport-config'
 import { CONSOLE_LOG_TRANSPORT_CONFIG } from './console-log-transport-config.injection-token'
 import { ConsoleLogTransport } from './console-log-transport.service'
 import { LogTransport } from '@shiftcode/logger'
+import { LoggerFeatureKind } from '../logger-feature-kind.enum'
+import { LoggerFeature } from '../logger-feature.type'
 
 export function withBrowserConsoleTransport(
   consoleLoggerConfigOrFactory: ConsoleLogTransportConfig | (() => ConsoleLogTransportConfig),
