+++ libs/core/src/lib/logger/logger-feature-kind.enum.ts
@@ -1,0 +1,5 @@
+/** @internal */
+export enum LoggerFeatureKind {
+  TRANSPORT,
+  REQUEST_INFO_FN,
+}
