+++ libs/core/src/lib/logger/logger-feature.type.ts
@@ -1,0 +1,7 @@
+import { Provider } from '@angular/core'
+import { LoggerFeatureKind } from './logger-feature-kind.enum'
+
+export interface LoggerFeature {
+  kind: LoggerFeatureKind
+  providers: Provider[]
+}
